buxfix for overflow in target calculation

+++ Src/main.c
@@ -569,7 +569,7 @@ int main(void)
 	  //throttle and PAS current target setting
 
 #if (DISPLAY_TYPE == DISPLAY_TYPE_BAFANG)
-	  uint16_mapped_PAS = map(uint32_PAS, RAMP_END, PAS_TIMEOUT, (PH_CURRENT_MAX*(int32_t)(assist_factor[ui8_AssistLevel]))>>8, 0); // level in range 0...5
+	  uint16_mapped_PAS = map(uint32_PAS, RAMP_END, PAS_TIMEOUT, ((int32_t)PH_CURRENT_MAX*(int32_t)(assist_factor[ui8_AssistLevel]))>>8, 0); // level in range 0...5
 #endif
 
 #if (DISPLAY_TYPE == DISPLAY_TYPE_KUNTENG)
@@ -581,7 +581,7 @@ int main(void)
 #endif
 
 #if (DISPLAY_TYPE == DISPLAY_TYPE_KINGMETER_901U)
-	  uint16_mapped_PAS = map(uint32_PAS, RAMP_END, PAS_TIMEOUT, (PH_CURRENT_MAX*(int32_t)(ui8_AssistLevel))>>8, 0); // level in range 0...255
+	  uint16_mapped_PAS = map(uint32_PAS, RAMP_END, PAS_TIMEOUT, ((int32_t)PH_CURRENT_MAX*(int32_t)(ui8_AssistLevel))>>8, 0); // level in range 0...255
 #endif
 
 #if (DISPLAY_TYPE == DISPLAY_TYPE_DEBUG)
